DB Completion - 2
Miiiiiiiight have not gotten to users yet, whatever...

Took me 2 days to properly sift through the checkout stuff, meaning checkout --> payments (INCOMPLETE) --> orders

So that should properly do checkout and orders (which can be accessed via the profile page too) BUT payments still needs to incorporate stripe, because that's a thing. Will do so next.

So far, orders are created and gotten correctly, no issues there. the API seems to be handling this well and a new DB View was created to handle ORDER complexity (vw_OrderDetails), which is obtaining the correct data.

As a side - the PhoneInput component is now fully functional, which is neat.

As another side - validation handles normalization as well now, which is also pretty neat.